Reuse 'is' function across fiber and test-renderer packages.
