+++ packages/shared/is.ts
