+++ packages/shared/tests/is.test.ts
@@ -1,8 +1,12 @@
-import { is } from '../../src/core/is'
+import { is } from '../is'
 
 describe('is', () => {
-  const myFunc = () => null
-  const myObj = { myProp: 'test-prop' }
+  const myFunc = () => {
+    return null
+  }
+  const myObj = {
+    myProp: 'test-prop',
+  }
   const myStr = 'test-string'
   const myNum = 1
   const myUnd = undefined
@@ -71,13 +75,8 @@ describe('is', () => {
 
     expect(is.equ('hello', 'hello')).toBe(true)
     expect(is.equ(1, 1)).toBe(true)
-    const obj = {
-      type: 'Mesh',
-    }
-    expect(is.equ(obj, obj)).toBe(true)
-
-    const arr = [1, 2, 3]
-    expect(is.equ(arr, arr)).toBe(true)
+    expect(is.equ(myObj, myObj)).toBe(true)
+    expect(is.equ(myArr, myArr)).toBe(true)
     expect(is.equ([1, 2, 3], [1, 2, 3])).toBe(true)
   })
 })
