--- packages/test-renderer/src/__tests__/is.test.ts
@@ -1,82 +1,0 @@
-import { is } from '../helpers/is'
-
-describe('is', () => {
-  const myFunc = () => {
-    return null
-  }
-  const myObj = {
-    myProp: 'test-prop',
-  }
-  const myStr = 'test-string'
-  const myNum = 1
-  const myUnd = undefined
-  const myArr = [1, 2, 3]
-
-  it('should tell me if something IS a function', () => {
-    expect(is.fun(myFunc)).toBe(true)
-
-    expect(is.fun(myObj)).toBe(false)
-    expect(is.fun(myStr)).toBe(false)
-    expect(is.fun(myNum)).toBe(false)
-    expect(is.fun(myUnd)).toBe(false)
-    expect(is.fun(myArr)).toBe(false)
-  })
-  it('should tell me if something IS an object', () => {
-    expect(is.obj(myFunc)).toBe(false)
-
-    expect(is.obj(myObj)).toBe(true)
-
-    expect(is.obj(myStr)).toBe(false)
-    expect(is.obj(myNum)).toBe(false)
-    expect(is.obj(myUnd)).toBe(false)
-    expect(is.obj(myArr)).toBe(false)
-  })
-  it('should tell me if something IS a string', () => {
-    expect(is.str(myFunc)).toBe(false)
-    expect(is.str(myObj)).toBe(false)
-
-    expect(is.str(myStr)).toBe(true)
-
-    expect(is.str(myNum)).toBe(false)
-    expect(is.str(myUnd)).toBe(false)
-    expect(is.str(myArr)).toBe(false)
-  })
-  it('should tell me if something IS a number', () => {
-    expect(is.num(myFunc)).toBe(false)
-    expect(is.num(myObj)).toBe(false)
-    expect(is.num(myStr)).toBe(false)
-
-    expect(is.num(myNum)).toBe(true)
-
-    expect(is.num(myUnd)).toBe(false)
-    expect(is.num(myArr)).toBe(false)
-  })
-  it('should tell me if something IS undefined', () => {
-    expect(is.und(myFunc)).toBe(false)
-    expect(is.und(myObj)).toBe(false)
-    expect(is.und(myStr)).toBe(false)
-    expect(is.und(myNum)).toBe(false)
-
-    expect(is.und(myUnd)).toBe(true)
-
-    expect(is.und(myArr)).toBe(false)
-  })
-  it('should tell me if something is an array', () => {
-    expect(is.arr(myFunc)).toBe(false)
-    expect(is.arr(myObj)).toBe(false)
-    expect(is.arr(myStr)).toBe(false)
-    expect(is.arr(myNum)).toBe(false)
-    expect(is.arr(myUnd)).toBe(false)
-
-    expect(is.arr(myArr)).toBe(true)
-  })
-  it('should tell me if something is equal', () => {
-    expect(is.equ([], '')).toBe(false)
-
-    expect(is.equ('hello', 'hello')).toBe(true)
-    expect(is.equ(1, 1)).toBe(true)
-    expect(is.equ(myObj, myObj)).toBe(true)
-    expect(is.equ(myArr, myArr)).toBe(true)
-    expect(is.equ([1, 2, 3], [1, 2, 3])).toBe(true)
-  })
-})
